Changed repo structure, requirements.txt and README to reflect move to venv
