+++ README.md
@@ -2,23 +2,36 @@
 An AI-powered assistant to answer questions about the contents of documents.
 
 ## Installation
-1. Make sure you have poetry installed on your system. If not, follow [these](https://python-poetry.org/docs/#installation) instructions to install it.
+1. Make sure you have python3 installed on your computer
 2. Clone the repo and move into it
     ```
     git clone https://github.com/Crown-Commercial-Service/ccs-ai-josh
     cd ccs-ai-josh
     ```
-3. Build the environment using poetry
+3. Create a virtual environment using venv, and activate it
     ```
-    poetry install
+    python -m venv venv
+    source venv/bin/activate
     ```
-4. Install the environment as a jupyter kernel
+4. Update pip in the new environment
     ```
-    poetry run python -m ipykernel install --user --name="ccs-ai-josh"
+    python -m pip install --upgrade pip
+    ```
+5. Install the dependencies into the environment
+    ```
+    python -m pip install -r requirements.txt
+    ```
+6. Install the repo as an internal module (so that internal imports work)
+    ```
+    python -m pip install -e .
+    ```
+7. Install the environment as a jupyter kernel
+    ```
+    python -m ipykernel install --user --name="ccs-ai-josh"
     ```
 5. Run the unit tests. If any fail, you may have an issue with your installation.
     ```
-    poetry run pytest ccs_ai_josh/tests/
+    pytest ccs_ai_josh/tests/
     ```
 Note that these tests require an internet connection to run, because they call an LLM.
 
@@ -27,8 +40,8 @@ Note that these tests require an internet connection to run, because they call a
 When contributing to this codebase, please follow these guidelines:
 1. The codebase uses a feature branch model, where new work is done on new feature branches created off the `develop` branch, and a pull request is submitted to merge the branch back into `develop`
 2. Please name feature branches using the convention TICKET-ID-short-description-of-work e.g. AI-91-basic-rag-system
-3. After creating a feature branch, make sure that your environment is up-to-date by running `poetry install`.
-4. If you need to install a new package for your contribution, please add it to the environment using `poetry add` and commit your new `poetry.lock` file along with your code.
+3. After creating a feature branch, make sure that your environment is up-to-date by running `python -m pip install -r requirements.txt`.
+4. If you need to install a new package for your contribution, please add it to the environment using `python -m pip freeze > requirements.txt` and commit your new `requirements.txt` file along with your code.
 5. Under **no** circumstances should **any** code be pushed directly to the `main` branch.
 
 ## Building the AI Josh System
@@ -41,7 +54,7 @@ To link to the documents that an answer is based on, we use a file called `CI_do
 ## Running the User Interface Locally
 If you want to run the user interface locally for development or testing purposes, run:
 ```
-poetry run streamlit run app.py
+streamlit run app.py
 ```
 
 ## Running Model Evaluation
@@ -57,7 +70,7 @@ Note that the full evaluation pipeline submits a total of 664 queries to an LLM 
 ## Exploring Model Evaluation Results
 The results of model evaluation can be explored in jupyter notebooks. To run the notebooks in this repo, launch a jupyter notebook server by running:
 ```
-poetry run jupyter lab
+jupyter lab
 ```
 In the jupyter lab landing page that launches in your browser, open one of the notebooks and select the kernel `ccs-ai-josh`.
