+++ prompts/groundedness_score_prompt.txt
