+++ pyproject.toml
@@ -1,45 +1,13 @@
-[project]
-name = "ccs-ai-josh"
-version = "0.1.0"
-description = ""
-authors = [
-    {name = "SamuelHLewis",email = "sam.lewis1@crowncommercial.gov.uk"}
-]
-readme = "README.md"
-requires-python = ">=3.13,<4.0"
-dependencies = [
-    "langchain[openai] (>=0.3.26,<0.4.0)",
-    "python-dotenv (>=1.1.1,<2.0.0)",
-    "langchain-community (>=0.3.27,<0.4.0)",
-    "azure-search-documents (>=11.5.3,<12.0.0)",
-    "azure-identity (>=1.23.0,<2.0.0)",
-    "asyncio (>=3.4.3,<4.0.0)",
-    "langgraph (>=0.5.1,<0.6.0)",
-    "streamlit (>=1.46.1,<2.0.0)",
-    "langchain-openai (>=0.3.28,<0.4.0)",
-    "openpyxl (>=3.1.5,<4.0.0)",
-    "seaborn (>=0.13.2,<0.14.0)",
-    "dvc (>=3.62.0,<4.0.0)",
-    "azure-storage-blob (>=12.26.0,<13.0.0)",
-    "reportlab (>=4.4.4,<5.0.0)",
-]
-
-
 [build-system]
-requires = ["poetry-core>=2.0.0,<3.0.0"]
-build-backend = "poetry.core.masonry.api"
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
 
-[tool.poetry]
-name = "ccs-ai-josh"
-version = "0.1.0"
-packages = [
-    { include = "src", from = "ccs_ai_josh" }
-]
-[tool.poetry.group.dev.dependencies]
-pytest = "^8.4.1"
-ipykernel = "^6.30.1"
-jupyterlab = "^4.4.9"
-jupyter = "^1.1.1"
-streamlit = "^1.50.0"
-seaborn = "^0.13.2"
+[project]
+name = "ccs_ai_josh"
+version = "0.0.1"
 
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
