+++ requirements.txt
@@ -1,80 +1,14 @@
-aiohappyeyeballs==2.6.1
-aiohttp-retry==2.9.1
-aiohttp==3.12.13
-aiosignal==1.4.0
-altair==5.5.0
-amqp==5.3.1
-annotated-types==0.7.0
-antlr4-python3-runtime==4.9.3
-anyio==4.9.0
-appdirs==1.4.4
-asyncio==3.4.3
-asyncssh==2.21.0
-atpublic==6.0.1
-attrs==25.3.0
 azure-common==1.1.28
 azure-core==1.35.0
 azure-identity==1.23.0
 azure-search-documents==11.5.3
-billiard==4.2.1
-blinker==1.9.0
-cachetools==6.1.0
-celery==5.5.3
-certifi==2025.6.15
-cffi==1.17.1
-charset-normalizer==3.4.2
-click-didyoumean==0.3.1
-click-plugins==1.1.1.2
-click-repl==0.3.0
-click==8.2.1
-colorama==0.4.6
-configobj==5.0.9
-contourpy==1.3.3
-cryptography==45.0.5
-cycler==0.12.1
-dataclasses-json==0.6.7
-dictdiffer==0.9.0
-diskcache==5.6.3
-distro==1.9.0
-dpath==2.2.0
-dulwich==0.24.1
-dvc-data==3.16.11
-dvc-http==2.32.0
-dvc-objects==5.1.1
-dvc-render==1.0.2
-dvc-studio-client==0.22.0
-dvc-task==0.40.2
-dvc==3.62.0
-entrypoints==0.4
-et-xmlfile==2.0.0
-filelock==3.18.0
-flatten-dict==0.4.2
-flufl-lock==8.2.0
-fonttools==4.59.0
-frozenlist==1.7.0
-fsspec==2025.7.0
-funcy==2.0
-gitdb==4.0.12
-gitpython==3.1.44
-grandalf==0.8
-greenlet==3.2.3
-gto==1.7.2
-h11==0.16.0
-httpcore==1.0.9
-httpx-sse==0.4.1
-httpx==0.28.1
-hydra-core==1.3.2
-idna==3.10
-isodate==0.7.2
-iterative-telemetry==0.0.10
-jinja2==3.1.6
-jiter==0.10.0
-jsonpatch==1.33
-jsonpointer==3.0.0
-jsonschema-specifications==2025.4.1
-jsonschema==4.24.0
-kiwisolver==1.4.8
-kombu==5.5.4
+azure-storage-blob==12.26.0
+beautifulsoup4
+dvc
+ipykernel
+ipython
+jinja2
+jupyterlab
 langchain-community==0.3.27
 langchain-core==0.3.68
 langchain-openai==0.3.28
@@ -85,88 +19,5 @@ langgraph-prebuilt==0.5.2
 langgraph-sdk==0.1.72
 langgraph==0.5.1
 langsmith==0.4.8
-markdown-it-py==4.0.0
-markupsafe==3.0.2
-marshmallow==3.26.1
-matplotlib==3.10.5
-mdurl==0.1.2
-msal-extensions==1.3.1
-msal==1.32.3
-multidict==6.6.3
-mypy-extensions==1.1.0
-narwhals==1.46.0
-networkx==3.5
-numpy==2.3.1
-omegaconf==2.3.0
-openai==1.93.1
-openpyxl==3.1.5
-orjson==3.10.18
-ormsgpack==1.10.0
-packaging==24.2
-pandas==2.3.1
-pathspec==0.12.1
-pillow==11.3.0
-platformdirs==4.3.8
-prompt-toolkit==3.0.51
-propcache==0.3.2
-protobuf==6.31.1
-psutil==7.0.0
-pyarrow==20.0.0
-pycparser==2.22
-pydantic-core==2.33.2
-pydantic-settings==2.10.1
-pydantic==2.11.7
-pydeck==0.9.1
-pydot==4.0.1
-pygit2==1.18.1
-pygments==2.19.2
-pygtrie==2.5.0
-pyjwt==2.10.1
-pyparsing==3.2.3
-python-dateutil==2.9.0.post0
-python-dotenv==1.1.1
-pytz==2025.2
-pywin32==311
-pyyaml==6.0.2
-referencing==0.36.2
-regex==2024.11.6
-requests-toolbelt==1.0.0
-requests==2.32.4
-rich==14.1.0
-rpds-py==0.26.0
-ruamel-yaml-clib==0.2.12
-ruamel-yaml==0.18.14
-scmrepo==3.5.2
-seaborn==0.13.2
-semver==3.0.4
-setuptools==80.9.0
-shellingham==1.5.4
-shortuuid==1.0.13
-shtab==1.7.2
-six==1.17.0
-smmap==5.0.2
-sniffio==1.3.1
-sqlalchemy==2.0.41
-sqltrie==0.11.2
-streamlit==1.46.1
-tabulate==0.9.0
-tenacity==9.1.2
-tiktoken==0.9.0
-toml==0.10.2
-tomlkit==0.13.3
-tornado==6.5.1
-tqdm==4.67.1
-typer==0.16.0
-typing-extensions==4.14.1
-typing-inspect==0.9.0
-typing-inspection==0.4.1
-tzdata==2025.2
-urllib3==2.5.0
-vine==5.1.0
-voluptuous==0.15.2
-watchdog==6.0.0
-wcwidth==0.2.13
-xxhash==3.5.0
-yarl==1.20.1
-zc-lockfile==3.0.post1
-zstandard==0.23.0
+matplotlib
+seaborn
