+++ scripts/basic_rag.py
@@ -1,17 +1,21 @@
-from src.llm_utils import check_index_naming
 import os
-# Azure vector store holds the vectors in a field called "text_vector", not "content_vector" as langchain expects
-os.environ["AZURESEARCH_FIELDS_CONTENT_VECTOR"] = "text_vector"
-# Azure vector store holds the document contents in a field called "chunk", not "content" as langchain expects
-os.environ["AZURESEARCH_FIELDS_CONTENT"] = "chunk"
 from dotenv import load_dotenv
 from azure.search.documents.indexes import SearchIndexClient
 from azure.core.credentials import AzureKeyCredential
 from langchain_community.vectorstores.azuresearch import AzureSearch
 from langchain_openai import AzureOpenAIEmbeddings, AzureChatOpenAI
-from src.multiturn_utils import build_graph, answer_once
+from typing_extensions import List, TypedDict
+from llm_utils import check_index_naming, generate_response
 
 load_dotenv()
+
+# before connecting to anything, check that the vector store fields are compatible with langchain
+index_client = SearchIndexClient(os.getenv("VECTOR_STORE_ENDPOINT"), AzureKeyCredential(os.getenv("VECTOR_STORE_KEY")))
+vector_store_name_status = check_index_naming(index_client=index_client, index_name=os.getenv("VECTOR_STORE_INDEX"))
+if vector_store_name_status:
+    print("Vector store is named correctly")
+else:
+    raise Exception("Vector store naming is not compatible with LangChain")
 
 embeddings: AzureOpenAIEmbeddings = AzureOpenAIEmbeddings(
     azure_deployment=os.getenv("EMBEDDING_DEPLOYMENT_NAME"),
@@ -39,13 +43,7 @@ llm = AzureChatOpenAI(
 )
 print("LLM connected")
 
-graph = build_graph(llm=llm, vector_store=vector_store)
-
-while True:
-    user_input = input("What do you want to know?\n")
-    response = answer_once(graph, user_input)
-    for i in range(len(response['source_contents'])):
-        print(f"###### Chunk {i+1} ######")
-        print(response['source_contents'][i])
-    print(response['source_names'])
-    print(response['answer'])
+while 1<2:
+    print("\n")
+    user_input = input("What do you want to know?\n\n")
+    print(generate_response(question=user_input, vector_store=vector_store, llm=llm))
