+++ scripts/extractCIDocsURLs.appscript
