+++ src/ccs_ai_josh.egg-info/PKG-INFO
@@ -1,0 +1,3 @@
+Metadata-Version: 2.4
+Name: ccs_ai_josh
+Version: 0.0.1
