+++ src/ccs_ai_josh.egg-info/SOURCES.txt
@@ -1,0 +1,8 @@
+README.md
+pyproject.toml
+src/ccs_ai_josh.egg-info/PKG-INFO
+src/ccs_ai_josh.egg-info/SOURCES.txt
+src/ccs_ai_josh.egg-info/dependency_links.txt
+src/ccs_ai_josh.egg-info/top_level.txt
+tests/test_eval_utils.py
+tests/test_llm_utils.py
